Desktop/fuelix/backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
sqlalchemy
sqlmodel
psycopg2-binary
alembic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
# For AI
numpy
pandas
email-validator
google-generativeai
